+++ api/reservation.js
@@ -1,3 +1,4 @@
+import { RESOURCE_TYPES, API_RETURN_MESSAGES, ITEM } from './constants.js';
 // POST /api/item/reserve/:id endpoint
 export async function reserveRoute(req, res,sql) {
