+++ api/sql-postgres.js
@@ -1,0 +1,11 @@
+import postgres from 'postgres';
+
+const sql = postgres({
+  host: process.env.DB_HOST,
+  port: process.env.DB_PORT,
+  database: process.env.DB_DATABASE,
+  username: process.env.DB_USERNAME,
+  password: process.env.DB_PASSWORD,
+});
+
+export default sql;
